requirements_flask.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.5
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.0
